Move out to fixtures.
